[ROSBE-UNIX] A corrective upload for 2.1.2 just has been performed.
ROSBE-68 patch was incorrectly applied and thus wasn't fixed in 2.1.2
Additionnaly, regarding ROSBE-72 has been added.

Regarding ROSBE-72, no extra efforts will be done for corrective upload for 2.1.2. Debian Stretch is a non released OS and its support with have to wait for 2.1.3 (or any later release) if it has been released in the meantime.

Sorry for the caused inconvenience regarding the broken upload of 2.1.2.

svn path=/trunk/RosBE/; revision=2245