Work to finally get to a new version.
Change all paths in NSIS to make creation of an installier more logical by just handing it a install root folder. Simplify the logic of the Shortcut creation to really work now.

Fix building of the platforn independent tools working on 64 Bit.

Dont hardcode the Host GCC to i386 anymore. This is needed to make a 64 Bit only setup for ROS 64 Bit.